+++ src/pylingdocs/cli.py
@@ -20,7 +20,7 @@ from pylingdocs.helpers import load_content
 from pylingdocs.helpers import new as create_new
 from pylingdocs.helpers import write_readme
 from pylingdocs.metadata import _load_metadata
-from pylingdocs.output import check_ids
+from pylingdocs.output import check_ids, check_abbrevs
 from pylingdocs.output import clean_output
 from pylingdocs.output import compile_latex as cmplatex
 from pylingdocs.output import create_output
@@ -156,6 +156,7 @@ def check(source, cldf, output_dir, latex):
         ),
     )
     check_ids(contents, ds, source)
+    check_abbrevs(ds, source)
 
 
 @main.command(cls=BuildCommand)